Needed to change how LimitedCluster worked, since the cluster-size-limit was set to 15. If the system had less than 15 chains, we ran into undefined behavior. Now we check if there are less than 15 chains; if so, we set the limit to tot_chains/2.
Made all the tot_chain-like entities size_t so there are no implicit conversions in mallocs

+++ src/initialize.h
@@ -19,16 +19,16 @@ float Temperature_Function(int mode, long nGen);
 
 void Calculate_Rot_Bias(float CurrentTemp);
 
-int *Create1DInt(const int xDim, const char* ArrName);
+int *Create1DInt(const size_t xDim, const char* ArrName);
 
-long *Create1DLong(const int xDim, const char* ArrName);
+long *Create1DLong(const size_t xDim, const char* ArrName);
 
-float *Create1DFloat(const int xDim, const char* ArrName);
+float *Create1DFloat(const size_t xDim, const char* ArrName);
 
-long double *Create1DLongdouble(const int xDim, const char* ArrName);
+long double *Create1DLongdouble(const size_t xDim, const char* ArrName);
 
-int **Create2DInt(const int xDim, const int yDim, const char* ArrName);
+int **Create2DInt(const size_t xDim, const size_t yDim, const char* ArrName);
 
-long double **Create2DLongdouble(const int xDim, const int yDim, const char* ArrName);
+long double **Create2DLongdouble(const size_t xDim, const size_t yDim, const char* ArrName);
 
 #endif // _INITIALIZE_H_
